% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_merge.R
\name{check_merge}
\alias{check_merge}
\title{Flag potential merge errors / mismatches.}
\usage{
check_merge(dataset, xVariable, yVariable, onlyMissingValues = FALSE)
}
\arguments{
\item{dataset}{The full dataset}

\item{xVariable}{The main dataset's column that the other dataset was merged
on (e.g. seed).}

\item{yVariable}{The column from the other dataset that you want to check
the merge on.}

\item{onlyMissingValues}{Boolean value (default = FALSE) to return only values
that are missing from the main dataset.}
}
\value{
Returns a dataframe with the missing values.
}
\description{
This function takes a dataset and checks for values by group. This
function was primarily designed to check merges against crop/seed and price
for p1/p2 trial analysis, but has other use cases as well.
}
\examples{


}
