% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commcare_gps_split.R
\name{commcare_gps_split}
\alias{commcare_gps_split}
\title{Split GPS data as saved by commcare for use in data.frame}
\usage{
commcare_gps_split(gps_col)
}
\arguments{
\item{gps_col}{the column in the df where the GPS character is saved.}
}
\description{
Split GPS data as saved by commcare for use in data.frame
}
\examples{
df <- cbind(df, commcare_gps_split(df$gps_col))
}
