\name{constrainArea-methods}
\docType{methods}
\alias{constrainArea-methods}
\alias{constrainArea}
\title{ Methods for Function constrainArea in Package `farmR' }
\description{
Constrain the area of a specific crop type
}
\section{Methods}{
\describe{

\item{farm = "FarmRepresentation", cropName = "character", lb = "numeric"}{ Constrain the area of a particular crop on a farm to be the value \code{lb}. If an optional \code{ub} argument is supplied the constraint is to hold the area of the crop between lb and ub }
}}
\keyword{methods}
