% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pairwise_means.R
\name{create_pairwise_means}
\alias{create_pairwise_means}
\title{Create table with pairwise means and confidence intervals}
\usage{
create_pairwise_means(depVar, testVar, dataset)
}
\arguments{
\item{dataset}{The full dataset}

\item{depVariable}{The name of the dependent variable from the dataset
(e.g. yield, profit)}

\item{testVariable}{The name of the variable that you are testing for
significant differences}
}
\value{
Returns a dataframe with the pairwise means and confidence intervals
for all trials.
}
\description{
This function takes a phase 1/phase 2 dataset, dependent variable, and
test variable and outputs a table showing the pairwise comparison of means
and confidence intervals for all possible combinations. It is designed
to match up with the Tukey's HSD output.
}
