% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tukey_output.R
\name{create_tukey_output}
\alias{create_tukey_output}
\title{Create Tukey HSD Output for RCBD Trials}
\usage{
create_tukey_output(completeFormula, depVar, testVar, dataset)
}
\arguments{
\item{completeFormula}{A complete regression formula to use for the
ANOVA test.}

\item{dataset}{The full dataset}

\item{depVariable}{The name of the dependent variable from the dataset
(e.g. yield, profit)}

\item{testVariable}{The name of the variable that you are testing for
significant differences}
}
\value{
Returns a dataframe with the pairwise comparison of all trials and
cleaned-p-values.
}
\description{
This function takes a clean formula, the already quosered dependent/test
variables, and full dataset to generate a cleaned-up dataframe of Tukey
HSD output.
}
