\name{cropArea-methods}
\docType{methods}
\alias{cropArea-methods}
\alias{cropArea}
\title{ Methods for Function cropArea in Package `farmR'}
\description{
 Methods for function \code{cropArea} in Package `farmR' 
}
\section{Methods}{
\describe{

\item{farm = "FarmRepresentation", cropName = "character"}{ Get the area cropped }
}}
\keyword{methods}
