% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farmr.R
\docType{package}
\name{farmr}
\alias{farmr}
\alias{farmr-package}
\title{farmr: A package with common analytical functions at 1AF.}
\description{
farmr: A package with common analytical functions at 1AF.
}
\section{Phase 1 Trial Functinos}{

The foo functions ...
}

