% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqr_check.R
\name{iqr_check}
\alias{iqr_check}
\title{Flag potential outliers using the IQR.}
\usage{
iqr_check(dataset, column, ...)
}
\arguments{
\item{dataset}{The full dataset}

\item{column}{The name of the variable to be checked for outliers.}

\item{....}{The names of the group (or groups) of data.
For example, trial, block, and/or AEZ. Note: groups must be in order with the
parent group (e.g. maize monocrop trial) before the child (e.g. the specific
variety).}
}
\value{
A vector of values with either a boolean value if the column is more
than 1.5 IQRs above the third quartile and below the first quartile.
}
\description{
This function takes a column of values and flags any values that are above a
set standard deviation from the mean. It is designed to help analysts flag
erroneous values.
}
\examples{
Without groups:
dataframe$Outlier <- sd_check(dataframe, yield)

With groups:
dataframe$Outlier <- iqr_check(dataframe, yield, block, trial, aez)

}
