% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_duplicated.R
\name{is_duplicated}
\alias{is_duplicated}
\title{Create a boolean flag to indicate the duplicated values in a variable.}
\usage{
is_duplicated(input_var, all_instances = TRUE)
}
\arguments{
\item{input_var}{The df column or vector to check}

\item{all_instances}{defaults to TRUE. This flag determines if all duplicates
should be flagged or only the last instance of the duplicate.}
}
\value{
A boolean vector of whether the value is duplicated in the vector or not
}
\description{
This function takes a vector or df column and flags whether a value is duplicated
in the vector. This simplifies finding all duplicated values in the data.
}
\examples{
is_duplicated(df$id)
is_duplicated(c(1:10, 1))
}
