% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_unique.R
\name{is_unique}
\alias{is_unique}
\title{Identify if a variable has only unique values}
\usage{
is_unique(input_var)
}
\arguments{
\item{input_var}{The df column or vector to check}
}
\value{
A boolean indicator of if it's uniquely identified or not.
}
\description{
This function takes a column from a df or vector and tells you if the variable
has only unique values
}
\examples{
is_unique(df$id)
is_unique(c(1,2,6,4,2))
}
