% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_report.R
\name{join_report}
\alias{join_report}
\title{This gives you a summary of what the joined data will look like following the sepcified join}
\usage{
join_report(.data, join_statement)
}
\arguments{
\item{join_statement}{the full dplyr join statement that creates the joined data}

\item{.datq}{The data in the dplyr pipe}
}
\description{
The problem it is aiming to solve is making it easy to see that
the join is working the way you expect it to.
}
\details{
The benefit of this approach is that it accepts all join statments. In an ideal world
we're able to easily see which ids are not matching based on what is provided to the
by feature in the command. That will be a future version of this function.
}
\examples{
join_report(band_members, left_join(band_members, band_instruments, by = "name"))
band_members \%>\% join_report(., left_join(., band_instruments, by = "name"))
}
