\name{model-methods}
\docType{methods}
\alias{model-methods}
\alias{model}
\title{ Methods for Function model in Package `farmR' }
\description{
Extract the internal java model object from a \code{FarmRepresentation} 
}
\section{Methods}{
\describe{

\item{farm = "FarmRepresentation"}{ Returns a java object reference holding the model }
}}
\keyword{methods}
