\name{objectiveScaleFactors-methods}
\docType{methods}
\alias{objectiveScaleFactors-methods}
\alias{objectiveScaleFactors}
\title{ Methods for Function objectiveScaleFactors in Package `farmR' }
\description{
Get the normalised objective weights
}
\section{Methods}{
\describe{

\item{farm = "FarmRepresentation"}{ Returns a list of normalised objective weights}
}}
\keyword{methods}
