\name{objectiveValues-methods}
\docType{methods}
\alias{objectiveValues-methods}
\alias{objectiveValues}
\title{ Methods for Function objectiveValues in Package `farmR' }
\description{
Get a list of objective values
}
\section{Methods}{
\describe{

\item{farm = "FarmRepresentation"}{ Get a list of objective values }
}}
\keyword{methods}
