% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plant_seed}
\alias{plant_seed}
\title{Plant a seed}
\usage{
plant_seed(project_name = NULL)
}
\arguments{
\item{project_name}{Name of the project (this names the project folder and
.Rprof file)}
}
\value{
Opens a new RStudio session with the newly created project
}
\description{
This function creates a R project folder, automatically creating a
subdirectory system I've found useful.
}
