% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_that_csv.R
\name{read_that_csv}
\alias{read_that_csv}
\title{Faster CSV Readings}
\usage{
read_that_csv(data)
}
\arguments{
\item{data}{Path to CSV file.}
}
\value{
dataframe object.
}
\description{
Faster CSV Readings
}
\examples{
data <- read_that_csv("data.csv")
}
