% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_ag_trials.R
\name{analyze_ag_trials}
\alias{analyze_ag_trials}
\title{Analyze Phase 1 and Phase 2 RCBD Agricultural Trials}
\usage{
analyze_ag_trials(dataset, trialVariable, depVariable, testVariable,
  profitVariable = NULL, ...)
}
\arguments{
\item{dataset}{The full dataset}

\item{trialVariable}{The name of the variable that designates which trial
you are referring to. This will be used to split the dataset.}

\item{depVariable}{The name of the dependent variable from the dataset
(e.g. yield, profit)}

\item{testVariable}{The name of the variable that you are testing for
significant differences}

\item{profitVariable}{The name of the variable that designates farmer profit.
If this is not included in your dataset, please set this variable to Null.}

\item{...}{The names of the other columns that the trial was stratified by /
source of variation to control for (e.g. block, AEZ)}
}
\value{
Returns a dataframe with the pairwise comparison of all trials.
}
\description{
This function takes a clean dataset, the dependent variable name, and the
names of the stratification variables and returns the analyzed output (using
Tukey's HSD) in a dataframe.
}
\examples{
With profit column:
rcbdOutput <- analyze_ag_trials(finalData, block_type_number, ton.hectare,
name, profitVariable = profit, block_number)

Without profit column:
rcbdOutput <- analyze_ag_trials(finalData, block_type_number, ton.hectare,
name, profitVariable = NULL, block_number)
}
